+++ sdk/storage/azure-storage-files-shares/CHANGELOG.md
@@ -15,6 +15,7 @@
 - Renamed `HasMorePages()` in paged response to `HasPage()`.
 - `ShareLeaseClient::Change()` updates internal lease id.
 - `ShareItem::ShareMetadata` was renamed to `ShareItem::Metadata`.
+- Renamed `ContentLength` in `FileItemDetails` to `FileSize`.
 
 ## 12.0.0-beta.10 (2021-04-16)