+++ sdk/storage/azure-storage-files-shares/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release History
 
+## 12.0.0-beta.12 (Unreleased)
+
+### Breaking Changes
+
+- Renamed `ContentLength` in `FileItemDetails` to `FileSize`.
+
+
 ## 12.0.0-beta.11 (2021-05-19)
 
 ### New Features